Fase 2/Evidencias Proyecto/Evidencias de sistema/Aplicación/Código sistema/3_FUNCIONES_AZURE/Funciones_azure_app/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pyodbc
scrypt
PyJWT
azure-communication-sms